[Build] - Updated puller and extractor logging - Added error checks and exit on download failures of packages to puller. - Updated setup-installer script to check exit status of package pull. - Updated build-installer script to check exit status of package extraction, UI build, and makeself build. - Add “noextract” arg to build-installer script to disable package extraction process of build.

+++ offline-installer-contained/setup-installer.sh
@@ -110,6 +110,12 @@ pushd package-puller
         exit 1
     fi
     
+    # check if package pull was successful
+    if [[ $? -ne 0 ]]; then
+        echo -e "\e[31mFailed pull of ROCm packages.\e[0m"
+        exit 1
+    fi
+    
     if [ -d $PULLER_OUTPUT ]; then
         echo -e "\e[93mExtraction directory exists. Removing: $PULLER_OUTPUT\e[0m"
         $SUDO rm -rf $PULLER_OUTPUT
@@ -141,6 +147,12 @@ pushd package-puller
         
     else
         echo Invalid Distro Type: $PULL_DISTRO_TYPE
+        exit 1
+    fi
+    
+    # check if package pull was successful
+    if [[ $? -ne 0 ]]; then
+        echo -e "\e[31mFailed pull of AMDGPU packages.\e[0m"
         exit 1
     fi
